+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/user_services_functions/start_user_services.go
@@ -555,6 +555,7 @@ func createStartServiceOperation(
 		user := serviceConfig.GetUser()
 		filesToBeMoved := serviceConfig.GetFilesToBeMoved()
 		tiniEnabled := serviceConfig.GetTiniEnabled()
+		ttyEnabled := serviceConfig.GetTtyEnabled()
 
 		// We replace the placeholder value with the actual private IP address
 		privateIPAddrStr := privateIpAddr.String()
@@ -731,6 +732,12 @@ func createStartServiceOperation(
 		}
 		if cmdArgs != nil {
 			createAndStartArgsBuilder.WithCmdArgs(cmdArgs)
+		}
+
+		if ttyEnabled {
+			// We don't care bout the height/width of the TTY. We just need it to be not nil.
+			defaultTtySize := &docker_manager.InteractiveModeTtySize{Height: 0, Width: 0}
+			createAndStartArgsBuilder.WithInteractiveModeTtySize(defaultTtySize)
 		}
 
 		createAndStartArgs := createAndStartArgsBuilder.Build()