+++ core/server/api_container/server/api_container_service.go
@@ -1188,6 +1188,7 @@ func getServiceInfoFromServiceObj(serviceObj *service.Service, serviceConfig *se
 		serviceConfig.GetNodeSelectors(),
 		serviceConfig.GetLabels(),
 		serviceConfig.GetTiniEnabled(),
+		serviceConfig.GetTtyEnabled(),
 	)
 
 	return serviceInfoResponse, nil