+++ CHANGELOG.md
@@ -1,5 +1,49 @@
 # Changelog
 
+## [1.10.3](https://github.com/kurtosis-tech/kurtosis/compare/1.10.2...1.10.3) (2025-07-17)
+
+
+### Bug Fixes
+
+* empty bridge network addr check ([#2770](https://github.com/kurtosis-tech/kurtosis/issues/2770)) ([4422ce8](https://github.com/kurtosis-tech/kurtosis/commit/4422ce8784c140d16c8642270700f5b222aabff4))
+* reduce logs agg container timeout ([#2777](https://github.com/kurtosis-tech/kurtosis/issues/2777)) ([1dc6ca4](https://github.com/kurtosis-tech/kurtosis/commit/1dc6ca424e810d4abe7b5507c32f3d0c7bd745e7))
+
+## [1.10.2](https://github.com/kurtosis-tech/kurtosis/compare/1.10.1...1.10.2) (2025-07-08)
+
+
+### Bug Fixes
+
+* docker auth bugs ([#2773](https://github.com/kurtosis-tech/kurtosis/issues/2773)) ([45ada05](https://github.com/kurtosis-tech/kurtosis/commit/45ada05521586ee6d2a1e3f48c1c3ae3a297c9e1))
+
+## [1.10.1](https://github.com/kurtosis-tech/kurtosis/compare/1.10.0...1.10.1) (2025-06-26)
+
+
+### Bug Fixes
+
+* refactor `HttpRequestService ` and `Recipe` to remove unnecessary calls to container engine ([#2768](https://github.com/kurtosis-tech/kurtosis/issues/2768)) ([8101cdc](https://github.com/kurtosis-tech/kurtosis/commit/8101cdcd9e5ec4c6e5eb2c94f8e040bdf6b51421))
+
+## [1.10.0](https://github.com/kurtosis-tech/kurtosis/compare/1.9.1...1.10.0) (2025-06-24)
+
+
+### Features
+
+* display instruction execution time for detailed verbosity ([#2767](https://github.com/kurtosis-tech/kurtosis/issues/2767)) ([e8e483e](https://github.com/kurtosis-tech/kurtosis/commit/e8e483ebb6f2cbf01efaaf19b5053b677d193ba5))
+
+## [1.9.1](https://github.com/kurtosis-tech/kurtosis/compare/1.9.0...1.9.1) (2025-06-19)
+
+
+### Bug Fixes
+
+* add podman mode to publish script ([#2765](https://github.com/kurtosis-tech/kurtosis/issues/2765)) ([b49d2f8](https://github.com/kurtosis-tech/kurtosis/commit/b49d2f8f328f066ca36111f9fd87196d285b0089))
+
+## [1.9.0](https://github.com/kurtosis-tech/kurtosis/compare/1.8.1...1.9.0) (2025-06-19)
+
+
+### Features
+
+* Add `tty_enabled` field to `ServiceConfig`. ([#2761](https://github.com/kurtosis-tech/kurtosis/issues/2761)) ([4313e31](https://github.com/kurtosis-tech/kurtosis/commit/4313e319e7e6333f2f4da4ac3bd28fb7e641ad4d))
+* support podman ([#2518](https://github.com/kurtosis-tech/kurtosis/issues/2518)) ([fe35eee](https://github.com/kurtosis-tech/kurtosis/commit/fe35eee5452cd7658955751fd01a6f3e7497e7d9))
+
 ## [1.8.1](https://github.com/kurtosis-tech/kurtosis/compare/1.8.0...1.8.1) (2025-05-21)